restructure frontend, update interfaces, revise fileupload and view, add memory-cache
