+++ package.json
@@ -35,6 +35,7 @@
     "@metamask/detect-provider": "^1.1.0",
     "@openzeppelin/contracts": "^2.5.0",
     "@types/bn.js": "^4.11.6",
+    "@types/memory-cache": "^0.2.1",
     "@types/react": "^16.9.23",
     "@types/react-modal": "^3.10.5",
     "@types/react-router-dom": "^5.1.5",
@@ -45,6 +46,7 @@
     "cors": "^2.8.5",
     "create-typescript-project": "^0.6.5",
     "ganache-cli": "^6.9.1",
+    "memory-cache": "^0.2.0",
     "node-sass": "^4.14.1",
     "openzeppelin-solidity": "^2.5.0",
     "react": "^16.13.0",
