+++ src/common/LetterContents.interface.ts
@@ -1,0 +1,7 @@
+import User from "./UserAuth.interface";
+export default interface Letter {
+    letterId: number;
+    contents: File;
+    requestor: User;
+    writer: User;
+}
