+++ src/common/LetterDetails.interface.ts
@@ -1,0 +1,6 @@
+import User from "./UserAuth.interface";
+export default interface LetterDetails {
+    letterId: number;
+    requestor: User;
+    writer: User;
+}
