+++ src/common/LetterHistory.interface.ts
@@ -1,0 +1,7 @@
+import User from "./UserAuth.interface";
+export default interface LetterHistory {
+    letterId: number;
+    requestor: User;
+    writer: User;
+    recipient: User;
+}
