+++ src/common/User.interface.ts
@@ -1,0 +1,4 @@
+export default interface User {
+  publicAddress: string,
+  name: string,
+}
