+++ src/common/UserAuth.interface.ts
@@ -1,0 +1,5 @@
+export default interface UserAuth {
+    publicAddress: string,
+    name: string,
+    jwtToken: string
+}
