+++ src/components/FileUpload/index.ts
@@ -1,0 +1,2 @@
+import FileUpload from "./FileUpload";
+export default FileUpload;
