+++ src/components/FileView/index.ts
@@ -1,0 +1,2 @@
+import FileView from "./FileView";
+export default FileView;
