--- src/interfaces/Letter.interface.ts
@@ -1,7 +1,0 @@
-import User from "./User.interface";
-export default interface Letter {
-    letter_id: number;
-    contents: File;
-    writer: User;
-    requestor: User;
-}
