--- src/interfaces/LetterCategory.enum.ts
@@ -1,5 +1,0 @@
-enum LetterCategory {
-  letters,
-  sentLetters,
-  invalid,
-} export default LetterCategory;
