--- src/interfaces/SentLetter.interface.ts
@@ -1,8 +1,0 @@
-import User from "./User.interface";
-export default interface SentLetter {
-    letter_id: number;
-    writer: User;
-    requestor: User;
-    recipient: User;
-    contents: File;
-}
