--- src/interfaces/User.interface.ts
@@ -1,6 +1,0 @@
-export default interface User {
-    publicAddress: string,
-    name: string,
-    email: string,
-    jwtToken: string,
-}
